backend/requirements.txt
# AssistantBot Backend Dependencies
# Optimized for Mac M3 (Metal acceleration)

# FastAPI & Web Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
sse-starlette>=1.8.0

# LLM & Embedding (M3 Metal optimized)
llama-cpp-python>=0.2.0
sentence-transformers>=2.3.0

# LangChain & RAG
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.0.0
beautifulsoup4>=4.12.0
pypdf>=3.17.0
